Mavenize project.
