+++ .gitignore
@@ -2,3 +2,4 @@ bin
 *\.class
 *\.mht
 *\.ser
+target
