+++ src/main/java/com/pessetto/CommandHandlers/Interfaces/Validatable.java
