+++ src/main/java/com/pessetto/Common/Variables.java
